feat!: delete the C app — EinkHome is Rust end to end
The C app (app/, 25 sources), its eh_lib FFI helper crate, and the C
lint tooling (clang-tidy/cppcheck/lizard glue) are gone.  The eh_ui
Rust workspace is the whole app:

- Makefile: cargo zigbuild targets — bookshelf.app (armel),
  bookshelf.armhf.app (armhf), bookshelf.pc (host SDL), and the new
  build/bookshelf.test (headless IPC host); lint = clippy + python.
  Clippy is clean workspace-wide (-D-warning grade; C-fidelity draws
  carry scoped too-many-arguments allows).
- sdk/build_armel.sh / build_armhf.sh + pb-demo/main.c stay: they link
  the eh_pb staticlib against the firmware rootfs.
- scripts: capture-visual builds armhf via make; run-visible-sdl doc
  points at make pc.
- CI: e2e workflow installs the Rust armv7 toolchain (zigbuild + zig,
  rust-cache); lint workflow swaps the C job for clippy.
- README rewritten around the Rust workspace layout.

Emulator e2e 52/52, SDL suites 40/40, unit tests 34/34.

+++ eh_ui/crates/eh_app/src/appui.rs
@@ -116,7 +116,7 @@ impl Widget for TopBar {
         // them): after the left button, after the source button, and the
         // left edges of the four right buttons.
         if !self.state.search {
-            ctx.vline((BTN_PAD + BTN_SIZE + 4) as u32, rect.y, rect.h, 2, col);
+            ctx.vline(BTN_PAD + BTN_SIZE + 4, rect.y, rect.h, 2, col);
             ctx.vline((SOURCE_BTN_X + SOURCE_BTN_W) as u32, rect.y, rect.h, 2, col);
             for k in 1..=4 {
                 let x = w - (BTN_PAD + k * BTN_SIZE) as i32;
